Renaming Histories to Approval. And some fixes.
