+++ src/modules/purchase/classes/queries/settlementApproval/ISettlementApprovalGetAllRequest.ts
@@ -1,4 +1,4 @@
-import { ISettlementApprovalGetAllFilter } from '@purchase/classes/filters/settlementHistories';
+import { ISettlementApprovalGetAllFilter } from '@purchase/classes/filters/settlementApproval';
 
 export interface ISettlementApprovalGetAllRequest {
   readonly filter: ISettlementApprovalGetAllFilter | undefined;
