+++ src/modules/purchase/classes/queries/settlementApproval/ISettlementApprovalPostRequest.ts
@@ -1,5 +1,5 @@
 import { IBaseCommand } from '@generic/interfaces';
-import { ISettlementApprovalPostPayload } from '@purchase/classes/request/settlementHistories';
+import { ISettlementApprovalPostPayload } from '@purchase/classes/request/settlementApproval';
 
 export interface ISettlementApprovalPostRequest extends IBaseCommand<ISettlementApprovalPostPayload> {
   companyUid: string;
