+++ src/modules/purchase/components/purchaseSettlement/editor/forms/PurchaseSettlementFormView.tsx
@@ -10,9 +10,7 @@ import { BaseFieldsProps, FieldArray, Fields, FormSection, WrappedFieldArrayProp
 export const PurchaseSettlementFormView: React.SFC<PurchaseSettlementFormProps> = props => {
   const {
     formMode,
-    formIsCurrencyIDR, formRate, 
-    // formActual,
-    // formDifference,
+    formIsCurrencyIDR, formRate,
     formActualValue,
     formDifferenceValue,
     change, initialValues
@@ -20,18 +18,6 @@ export const PurchaseSettlementFormView: React.SFC<PurchaseSettlementFormProps> 
 
   const fields = Object.getOwnPropertyNames(initialValues.information);
 
-  // const onChangeValueIdr = (event: any, newValue: number) => {
-  //   change('information.actualInIDR', newValue * formRate);
-  //   change('information.differenceInIDR', newValue * formRate );
-  // };
-
-  // const onChangeActual = (event: any, newValue: number, oldValue: number) => {
-  //   change('information.actual', formActual);
-  // };
-
-  // const onChangeDifference = (event: any, newValue: number, oldValue: number) => {
-  //   change('information.difference', formDifference );
-  // };
   const onChangeValueActual = (event: any, newValue: number, oldValue: number) => {
     change('information.actualInIDR', formActualValue * formRate );
   };
@@ -45,9 +31,6 @@ export const PurchaseSettlementFormView: React.SFC<PurchaseSettlementFormProps> 
       formMode={formMode}
       context={context}
       isCurrencyIdr={formIsCurrencyIDR}
-      // onChangeValueIdr={onChangeValueIdr}
-      // onChangeActual={onChangeActual}
-      // onChangeDifference={onChangeDifference}
       onChangeActualValue={onChangeValueActual}
       onChangeDifferenceValue={onChangeValueDifference}
     />
