+++ src/modules/purchase/store/reducers/index.ts
@@ -7,9 +7,9 @@ export * from './purchaseSettlement/settlementGetAllReducer';
 export * from './purchaseSettlement/settlementGetByIdReducer';
 export * from './purchaseSettlement/settlementPostReducer';
 export * from './purchaseSettlement/settlementPutReducer';
-export * from './purchaseHistories/purchaseApprovalGetAllReducer';
-export * from './purchaseHistories/purchaseApprovalGetByIdReducer';
-export * from './purchaseHistories/purchaseApprovalPostReducer';
-export * from './settlementHistories/settlementApprovalGetAllReducer';
-export * from './settlementHistories/settlementApprovalGetByIdReducer';
-export * from './settlementHistories/settlementApprovalPostReducer';
+export * from './purchaseApproval/purchaseApprovalGetAllReducer';
+export * from './purchaseApproval/purchaseApprovalGetByIdReducer';
+export * from './purchaseApproval/purchaseApprovalPostReducer';
+export * from './settlementApproval/settlementApprovalGetAllReducer';
+export * from './settlementApproval/settlementApprovalGetByIdReducer';
+export * from './settlementApproval/settlementApprovalPostReducer';
