+++ src/modules/purchase/store/reducers/purchaseApproval/purchaseApprovalGetAllReducer.ts
@@ -1,5 +1,5 @@
 import { IQueryCollectionState } from '@generic/interfaces';
-import { IPurchaseApprovalGetAllRequest } from '@purchase/classes/queries/purchaseHistories';
+import { IPurchaseApprovalGetAllRequest } from '@purchase/classes/queries/purchaseApproval';
 import { IPurchase } from '@purchase/classes/response/purchaseRequest';
 import { PurchaseApprovalAction as Action } from '@purchase/store/actions';
 import { Reducer } from 'redux';
